fix dirichlet sample memory leak
