+++ paddle/phi/kernels/cpu/dirichlet_kernel.cc
@@ -17,7 +17,6 @@
 #include "paddle/phi/core/kernel_registry.h"
 #include "paddle/phi/kernels/cpu/elementwise.h"
 #include "paddle/phi/kernels/cpu/reduce.h"
-#include "paddle/phi/kernels/funcs/dirichlet.h"
 #include "paddle/phi/kernels/funcs/elementwise_functor.h"
 #include "paddle/phi/kernels/funcs/for_range.h"
 #include "paddle/phi/kernels/funcs/reduce_functor.h"
@@ -29,20 +28,20 @@ template <typename T, typename UniformSamplerT, typename NormalSamplerT>
 struct GammaCPUFunctor {
   GammaCPUFunctor(const T* alpha,
                   T* gamma,
-                  funcs::BaseSampler<T, UniformSamplerT> uniform,
-                  funcs::BaseSampler<T, NormalSamplerT> normal)
+                  BaseSampler<T, UniformSamplerT> uniform,
+                  BaseSampler<T, NormalSamplerT> normal)
       : alpha_(alpha), gamma_(gamma), uniform_(uniform), normal_(normal) {}
 
   HOST void operator()(int64_t index) {
-    auto sample = funcs::sample_gamma<T, T, UniformSamplerT, NormalSamplerT>(
+    auto sample = sample_gamma<T, T, UniformSamplerT, NormalSamplerT>(
         alpha_[index], uniform_, normal_);
     gamma_[index] = std::max(std::numeric_limits<T>::min(), sample);
   }
 
   const T* alpha_;
   T* gamma_;
-  funcs::BaseSampler<T, UniformSamplerT> uniform_;
-  funcs::BaseSampler<T, NormalSamplerT> normal_;
+  BaseSampler<T, UniformSamplerT> uniform_;
+  BaseSampler<T, NormalSamplerT> normal_;
 };
 
 template <typename T>
@@ -56,49 +55,44 @@ struct DirichletSampler<CPUContext, T> {
       std::uniform_real_distribution<T> u(0.0, 1.0);
       return u(*generator);
     };
-    funcs::BaseSampler<T, decltype(uniform)> standard_uniform(uniform);
+    BaseSampler<T, decltype(uniform)> standard_uniform(uniform);
 
     auto normal = [&generator]() {
       std::normal_distribution<T> n(0.0, 1.0);
       return n(*generator);
     };
-    funcs::BaseSampler<T, decltype(normal)> standard_normal(normal);
+    BaseSampler<T, decltype(normal)> standard_normal(normal);
 
     // sample from K gamma distributions, where K=alpha.numel()
-    DenseTensor* gamma_samples = new DenseTensor();
-    gamma_samples->Resize(alpha.dims());
-    dev_ctx.template Alloc<T>(gamma_samples);
+    DenseTensor gamma_samples;
+    gamma_samples.Resize(alpha.dims());
+    dev_ctx.template Alloc<T>(&gamma_samples);
 
     GammaCPUFunctor<T, decltype(uniform), decltype(normal)> gamma_functor(
         alpha.data<T>(),
-        gamma_samples->data<T>(),
+        gamma_samples.data<T>(),
         standard_uniform,
         standard_normal);
     funcs::ForRange<CPUContext> for_range(dev_ctx, alpha.numel());
     for_range(gamma_functor);
 
     // normalize them into a simplex, along the last axis
-    DenseTensor* gamma_sum = new DenseTensor();
-    auto new_shape = gamma_samples->dims();
+    DenseTensor gamma_sum;
+    auto new_shape = gamma_samples.dims();
     new_shape[new_shape.size() - 1] = 1;
-    gamma_sum->Resize(new_shape);
-    dev_ctx.template Alloc<T>(gamma_sum);
+    gamma_sum.Resize(new_shape);
+    dev_ctx.template Alloc<T>(&gamma_sum);
 
     ReduceKernelImpl<CPUContext, T, T, funcs::SumFunctor>(
         dev_ctx,
-        *gamma_samples,
-        gamma_sum,
+        gamma_samples,
+        &gamma_sum,
         {new_shape.size() - 1},
         true,
         false);
 
     funcs::ElementwiseCompute<funcs::DivideFunctor<T>, T, T>(
-        dev_ctx,
-        *gamma_samples,
-        *gamma_sum,
-        -1,
-        funcs::DivideFunctor<T>(),
-        out);
+        dev_ctx, gamma_samples, gamma_sum, -1, funcs::DivideFunctor<T>(), out);
   }
 };
